Allow orga managers to change team leaders in their teams

+++ app/controllers/organizations/managers_controller.rb
@@ -1,13 +1,15 @@
 class Organizations::ManagersController < ApplicationController
 
-  before_action :require_admin
+  before_action :find_organization_by_id
+  before_action :require_admin_or_manager
 
   def update
-    @organization = Organization.find(params[:id])
     # Managers
-    @managers = User.active.where(id: params[:manager_ids])
-    OrganizationManager.where(user_id: params[:manager_ids]).delete_all
-    @organization.managers = @managers
+    if User.current.admin? # Managers are not allowed to modify managers
+      @managers = User.active.where(id: params[:manager_ids])
+      OrganizationManager.where(user_id: params[:manager_ids]).delete_all
+      @organization.managers = @managers
+    end
     # Team leaders
     @team_leaders = User.active.where(id: params[:team_leader_ids])
     OrganizationTeamLeader.where(user_id: params[:team_leader_ids]).delete_all
